server/main/requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
black==20.8b1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
cssutils==2.2.0
ecdsa==0.14.1
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
idna==2.10
jose==1.0.0
lxml==4.6.3
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.8.1
premailer==3.8.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
PyMySQL==1.0.2
python-dotenv==0.17.0
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.9
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
yagmail==0.14.247
